+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsFastInitBoxTestGenerated.java
@@ -1838,6 +1838,12 @@ public class JsFastInitBoxTestGenerated extends AbstractJsFastInitBoxTest {
       }
 
       @Test
+      @TestMetadata("SwitchingSuspendProvidersWithShards.kt")
+      public void testSwitchingSuspendProvidersWithShards() {
+        run("SwitchingSuspendProvidersWithShards.kt");
+      }
+
+      @Test
       @TestMetadata("ThreeShardsWithOrder.kt")
       public void testThreeShardsWithOrder() {
         run("ThreeShardsWithOrder.kt");